Execute event bus inbound interceptor before delivering the message rather than at scheduling time
